feat: implement medium priority improvements (v0.5.0)
### Modular Architecture Implementation
- Split large import-manager.ts (732 lines) into focused modules in src/lib/import/
- Split data-validator.ts (694 lines) into specialized modules in src/lib/validation/
- Split IdeasTableRow.tsx (351 lines) into three focused components in src/components/ideas/
- Maintained full backwards compatibility through re-export patterns

### React Query Data Management
- Added @tanstack/react-query dependency for intelligent caching and optimistic updates
- Implemented comprehensive useMealsQuery and useIdeasQuery hooks
- Added mutation hooks for all CRUD operations with error recovery
- Enhanced offline-first architecture with smart cache invalidation
- Integrated query client into _app.tsx with proper configuration

### Performance & Code Quality
- Improved component modularity and single responsibility principle
- Enhanced TypeScript interface segregation
- Added memoization and optimistic updates for better UX
- Reduced code complexity through focused module separation

### Version Management
- Updated package.json, index.tsx, and export-manager.ts to v0.5.0
- Comprehensive CHANGELOG.md entry documenting all improvements
