+++ src/lib/export-manager.ts
@@ -65,7 +65,7 @@ export interface SerializableMeal {
 }
 
 export class ExportManager {
-  private static readonly VERSION = '0.4.3';
+  private static readonly VERSION = '0.5.0';
   private static readonly SOURCE = 'dish-diary';
 
   /**
